4. Тесты для стороны прямоугольника. [#418184]

+++ src/ru/job4j/condition/SqArea.java
@@ -1,10 +1,10 @@
 package ru.job4j.condition;
 
 public class SqArea {
-    public static double square(int p, int k) {
-        int h = p/(2*(k+1));
-        int l = h*k;
-        double rsl = l*h;
+    public static double square(double p, double k) {
+        double h = p / (2 * (k + 1));
+        double l = h * k;
+        double rsl = l * h;
         return rsl;
     }
